+++ docs/changelog.md
@@ -70,6 +70,7 @@ Added:
 - [TOFAnalogResponse][extra.applications.TOFAnalogResponse] can deconvolve the Auger-Meitner
   and photon line if the flag `deconvolve` is set to True.
   summary about a pulse pattern (!435).
+- `StepTimer` class for measuring performance within functions.
 
 Fixed: